[CHANGED] Refactored all data-types for cycle and period references to uint64 and uint32, respectively

+++ contracts/mocks/staking/NftStakingMock.sol
@@ -10,7 +10,7 @@ contract NftStakingMock is NftStaking {
 
     constructor(
         uint256 cycleLength_,
-        uint256 payoutPeriodLength_,
+        uint32 payoutPeriodLength_,
         uint256 freezeDurationAfterStake_,
         address whitelistedNftContract_,
         address rewardsToken_,
